requirements.txt
pandas
python-dotenv
psutil
mysql-connector-python
sqlalchemy
reportlab
